Update Help_Menu.adoc (#565)
The link to the tutorials is incorrect; I wrote the link that I thought it should be, please check if it is correct so I can also fix it in the Spanish version. In the same paragraph, it said "GeoGebraWiki" as one word, and I added a space that seemed correct to me.