Update Help_Menu.adoc (#566)
* Update Help_Menu.adoc

The link to the tutorials is incorrect; I wrote the link that I thought it should be, please check if it is correct so I can also fix it in the Spanish version. In the same paragraph, it said "GeoGebraWiki" as one word, and I added a space that seemed correct to me.

* Update Help_Menu.adoc

* The images are missing in the titles; I couldn’t find them in the folders.
* The note that appears at the beginning was incorrectly linked (with the hyperlink).
* The title "Manual" was missing.
* The Help menu did not lead to the Help Center, but to the manual.
* The GeoGebra forum is no longer in this menu, and I chose to remove it; please check.